Time: 1 ms (99.5%), Space: 41.4 MB (98.89%) - LeetHub

+++ 3379-score-of-a-string/3379-score-of-a-string.java
@@ -3,7 +3,7 @@ class Solution {
         int sum=0;
        for(int i=0;i<s.length()-1;i++)
        {
-           sum += Math.abs(s.charAt(i)-s.charAt(i+1));
+           sum +=(int) Math.abs(s.charAt(i)-s.charAt(i+1));
        } 
        return sum;
     }
